+++ lib/minimap2/kthread.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif
 
-void kt_for(int n_threads, void (*func)(void*,long,int), void *data, long n);
+void kt_for(int n_threads, void (*func)(void*,long long,int), void *data, long long n);
 void kt_pipeline(int n_threads, void *(*func)(void*, int, void*), void *shared_data, int n_steps);
 
 #ifdef __cplusplus
